+++ backend/routes/customers.js
@@ -32,10 +32,11 @@ router.post(
         telephone,
         address,
         notes,
-        lead_source
+        lead_source,
+        created_by
       )
       VALUES
-      ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15)
+      ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,$15,$16)
       RETURNING *`,
       [
         r.Company,
@@ -53,6 +54,7 @@ router.post(
         r.Address,
         r.Notes,
         r.Source,
+        req.user.userId,
       ]
     );